+++ tests/testing/vcr/integration/test_sync_client.py
@@ -24,8 +24,19 @@ import pytest
 
 from tchannel.errors import UnexpectedError
 from tchannel.sync import TChannel
-from tchannel.thrift import thrift_request_builder
 from tchannel.testing import vcr
+from tchannel import thrift
+
+
+@pytest.fixture
+def thrift_module(mock_server):
+    # TODO dont dup between async and sync
+    service = thrift.load(
+        path='tests/data/idls/ThriftTest2.thrift',
+        service='myservice',
+        hostport=mock_server.hostport,
+    )
+    return service
 
 
 @pytest.mark.gen_test
@@ -154,23 +165,23 @@ def test_record_success_new_channels(tmpdir, mock_server):
 
 
 @pytest.mark.gen_test
-def test_record_success_thrift(tmpdir, thrift_service, mock_server):
-    myservice = thrift_request_builder(
-        'myservice', thrift_service, hostport=mock_server.hostport
-    )
-
+def test_record_success_thrift(tmpdir, thrift_module, mock_server):
     path = tmpdir.join('data.yaml')
-    expected_item = thrift_service.Item(
-        'foo', thrift_service.Value(stringValue='bar')
+    expected_item = thrift_module.Item(
+        'foo', thrift_module.Value(stringValue='bar')
     )
-    mock_server.expect_call(thrift_service, method='getItem').and_result(
+    mock_server.expect_call(
+        thrift_module.Service, method='getItem'
+    ).and_result(
         expected_item
     ).once()
 
-    client = TChannel('test')
+    tchannel = TChannel('test')
 
     with vcr.use_cassette(str(path)) as cass:
-        response = client.thrift(myservice.getItem('foo')).result(1)
+        response = tchannel.thrift(
+            thrift_module.Service.getItem('foo')
+        ).result(1)
         item = response.body
         assert item == expected_item
 
@@ -178,7 +189,9 @@ def test_record_success_thrift(tmpdir, thrift_service, mock_server):
     assert path.check(file=True)
 
     with vcr.use_cassette(str(path)) as cass:
-        response = client.thrift(myservice.getItem('foo')).result(1)
+        response = tchannel.thrift(
+            thrift_module.Service.getItem('foo')
+        ).result(1)
         item = response.body
         assert item == expected_item
 
@@ -207,28 +220,30 @@ def test_protocol_exception(tmpdir, mock_server):
 
 
 @pytest.mark.gen_test
-def test_record_thrift_exception(tmpdir, mock_server, thrift_service):
+def test_record_thrift_exception(tmpdir, mock_server, thrift_module):
     path = tmpdir.join('data.yaml')
 
-    myservice = thrift_request_builder(
-        'myservice', thrift_service, hostport=mock_server.hostport
-    )
-
-    mock_server.expect_call(thrift_service, method='getItem').and_raise(
-        thrift_service.ItemDoesNotExist('foo')
+    mock_server.expect_call(
+        thrift_module.Service, method='getItem'
+    ).and_raise(
+        thrift_module.ItemDoesNotExist('foo')
     ).once()
 
-    client = TChannel('test')
+    tchannel = TChannel('test')
 
     with vcr.use_cassette(str(path)) as cass:
-        with pytest.raises(thrift_service.ItemDoesNotExist):
-            client.thrift(myservice.getItem('foo')).result(1)
+        with pytest.raises(thrift_module.ItemDoesNotExist):
+            tchannel.thrift(
+                thrift_module.Service.getItem('foo')
+            ).result(1)
 
     assert cass.play_count == 0
     assert path.check(file=True)
 
     with vcr.use_cassette(str(path)) as cass:
-        with pytest.raises(thrift_service.ItemDoesNotExist):
-            client.thrift(myservice.getItem('foo')).result(1)
+        with pytest.raises(thrift_module.ItemDoesNotExist):
+            tchannel.thrift(
+                thrift_module.Service.getItem('foo')
+            ).result(1)
 
     assert cass.play_count == 1